requirements.txt
Babel==2.3.4
Django==1.9.6
Markdown==2.6.7
Pygments==2.1.3
appnope==0.1.0
backports.shutil-get-terminal-size==1.0.0
boto==2.43.0
colorlog==2.10.0
coreapi==2.0.9
decorator==4.0.10
dj-database-url==0.4.1
django-filter==0.15.3
django-phonenumber-field==1.1.0
django-rest-swagger==2.1.0
django-storages==1.5.1
django-twitter-bootstrap==3.3.0
djangorestframework==3.5.3
enum34==1.1.6
gevent==1.1.2
greenlet==0.4.10
gunicorn==19.6.0
ipython==5.1.0
ipython-genutils==0.1.0
itypes==1.1.0
openapi-codec==1.1.7
pathlib2==2.1.0
pexpect==4.2.1
phonenumberslite==7.7.4
pickleshare==0.7.4
prompt-toolkit==1.0.9
psycopg2==2.6.2
ptyprocess==0.5.1
pytz==2016.7
requests==2.12.1
simplegeneric==0.8.1
simplejson==3.10.0
six==1.10.0
traitlets==4.3.1
uritemplate==3.0.0
wcwidth==0.1.7
wsgiref==0.1.2
